requirements.txt
blobfile==3.0.0
filelock==3.18.0
fsspec==2025.5.1
-e git+https://github.com/RandyBrown1965/GenerativeDiffusionPrior.git@b34494cd53e344a5c726e502f552cd7fb888aad0#egg=guided_diffusion
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.12.0.88
pillow==11.3.0
pycryptodomex==3.23.0
pytorch_ssim==0.1
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.14.1
urllib3==2.5.0
